+++ sketch-core/src/commonTest/kotlin/com/github/panpf/sketch/core/common/test/state/ThumbnailMemoryCacheStateImageTest.kt
@@ -16,24 +16,28 @@
 
 package com.github.panpf.sketch.core.common.test.state
 
+import com.github.panpf.sketch.asImage
+import com.github.panpf.sketch.cache.ImageCacheValue
+import com.github.panpf.sketch.cache.newCacheValueExtras
+import com.github.panpf.sketch.decode.ImageInfo
+import com.github.panpf.sketch.decode.internal.createInSampledTransformed
+import com.github.panpf.sketch.decode.internal.createResizeTransformed
 import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.request.ImageRequest
-import com.github.panpf.sketch.resize.Precision.EXACTLY
 import com.github.panpf.sketch.resize.Precision.LESS_PIXELS
+import com.github.panpf.sketch.resize.Resize
+import com.github.panpf.sketch.resize.Scale
 import com.github.panpf.sketch.state.ThumbnailMemoryCacheStateImage
 import com.github.panpf.sketch.test.singleton.getTestContextAndSketch
 import com.github.panpf.sketch.test.utils.FakeImage
 import com.github.panpf.sketch.test.utils.FakeStateImage
-import com.github.panpf.sketch.test.utils.TestCountTarget
+import com.github.panpf.sketch.test.utils.createBitmap
 import com.github.panpf.sketch.test.utils.getTestContext
 import com.github.panpf.sketch.test.utils.toRequestContext
-import com.github.panpf.sketch.transform.CircleCropTransformation
-import com.github.panpf.sketch.transform.RotateTransformation
-import com.github.panpf.sketch.transform.RoundedCornersTransformation
+import com.github.panpf.sketch.util.Size
 import com.github.panpf.sketch.util.SketchSize
-import kotlinx.coroutines.Dispatchers
+import com.github.panpf.sketch.util.div
 import kotlinx.coroutines.test.runTest
-import kotlinx.coroutines.withContext
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
@@ -41,6 +45,30 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 
 class ThumbnailMemoryCacheStateImageTest {
+
+    @Test
+    fun testConstructor() {
+        ThumbnailMemoryCacheStateImage().apply {
+            assertNull(uri)
+            assertNull(defaultImage)
+            assertEquals(-1, maxMismatchCount)
+        }
+        ThumbnailMemoryCacheStateImage("uri").apply {
+            assertEquals("uri", uri)
+            assertNull(defaultImage)
+            assertEquals(-1, maxMismatchCount)
+        }
+        ThumbnailMemoryCacheStateImage("uri", FakeStateImage()).apply {
+            assertEquals("uri", uri)
+            assertNotNull(defaultImage)
+            assertEquals(-1, maxMismatchCount)
+        }
+        ThumbnailMemoryCacheStateImage("uri", FakeStateImage(), 10).apply {
+            assertEquals("uri", uri)
+            assertNotNull(defaultImage)
+            assertEquals(10, maxMismatchCount)
+        }
+    }
 
     @Test
     fun testKey() {
@@ -51,20 +79,20 @@ class ThumbnailMemoryCacheStateImageTest {
         val defaultImage = FakeStateImage()
         ThumbnailMemoryCacheStateImage(uri, defaultImage).apply {
             assertEquals(
-                expected = "ThumbnailMemoryCache('$uri',${defaultImage.key})",
+                expected = "ThumbnailMemoryCache('$uri',${defaultImage.key},-1)",
                 actual = key
             )
         }
         val defaultImage1 = FakeStateImage(FakeImage(SketchSize(200, 200)))
-        ThumbnailMemoryCacheStateImage(uri, defaultImage1).apply {
+        ThumbnailMemoryCacheStateImage(uri, defaultImage1, 10).apply {
             assertEquals(
-                expected = "ThumbnailMemoryCache('$uri',${defaultImage1.key})",
+                expected = "ThumbnailMemoryCache('$uri',${defaultImage1.key},10)",
                 actual = key
             )
         }
         ThumbnailMemoryCacheStateImage(null, null).apply {
             assertEquals(
-                expected = "ThumbnailMemoryCache(null,null)",
+                expected = "ThumbnailMemoryCache(null,null,-1)",
                 actual = key
             )
         }
@@ -74,274 +102,263 @@ class ThumbnailMemoryCacheStateImageTest {
     fun testGetImage() = runTest {
         val (context, sketch) = getTestContextAndSketch()
 
+        val imageFile = ResourceImages.jpeg
+        val request = ImageRequest(context, imageFile.uri)
+        val requestContext = request.toRequestContext(sketch, Size(1080, 1920))
+        val thumbnailSize = imageFile.size / 8f
+        val thumbnailImage = createBitmap(thumbnailSize.width, thumbnailSize.height).asImage()
+        val imageInfo = ImageInfo(imageFile.size, "image/jpeg")
+        val resize = Resize(500, 500, LESS_PIXELS, Scale.CENTER_CROP)
+
         val memoryCache = sketch.memoryCache
         memoryCache.clear()
         assertEquals(0, memoryCache.keys().size)
 
-        val requests1 = arrayOf(
-            ImageRequest(context, ResourceImages.jpeg.uri) {
-                size(100, 100)
-                precision(LESS_PIXELS)
-                target(TestCountTarget())
-            },
-            ImageRequest(context, ResourceImages.jpeg.uri) {
-                size(100, 100)
-                precision(EXACTLY)
-                target(TestCountTarget())
-            },
-            ImageRequest(context, ResourceImages.jpeg.uri) {
-                size(100, 100)
-                precision(LESS_PIXELS)
-                target(TestCountTarget())
-                transformations(CircleCropTransformation())
-            },
-        )
-        val requests2 = arrayOf(
-            ImageRequest(context, ResourceImages.png.uri) {
-                size(100, 100)
-                precision(LESS_PIXELS)
-                target(TestCountTarget())
-            },
-            ImageRequest(context, ResourceImages.png.uri) {
-                size(100, 100)
-                precision(EXACTLY)
-                target(TestCountTarget())
-            },
-            ImageRequest(context, ResourceImages.png.uri) {
-                size(100, 100)
-                precision(LESS_PIXELS)
-                target(TestCountTarget())
-                transformations(CircleCropTransformation())
-            },
-        )
-
+        // empty
         assertEquals(
-            6,
-            requests1.map { it.toRequestContext(sketch).memoryCacheKey }
-                .plus(requests2.map { it.toRequestContext(sketch).memoryCacheKey }).distinct().size
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
         )
 
-        withContext(Dispatchers.Main) {
-            val inexactlyStateImage = ThumbnailMemoryCacheStateImage()
-            val inexactlyStateImage1 = ThumbnailMemoryCacheStateImage(requests1[0].uri.toString())
-            val inexactlyStateImage2 = ThumbnailMemoryCacheStateImage(requests2[0].uri.toString())
+        // defaultImage
+        val defaultImage = FakeStateImage()
+        assertEquals(
+            expected = defaultImage.getImage(sketch, request, null),
+            actual = ThumbnailMemoryCacheStateImage(defaultImage = defaultImage)
+                .getImage(sketch, request, null)
+        )
 
-            assertEquals(0, memoryCache.keys().size)
-            requests1.plus(requests2).forEach { request ->
-                assertNull(
-                    memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                    request.toRequestContext(sketch).memoryCacheKey,
+        // default uri
+        memoryCache.put(
+            key = requestContext.memoryCacheKey,
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(imageInfo, resize, transformeds = null, extras = null)
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = thumbnailImage,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // custom uri and startsWith error
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage(
+                uri = imageFile.uri.replace("://", ".custom://")
+            ).getImage(sketch, request, null)
+        )
+
+        // memoryCacheKey.length error
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        memoryCache.put(
+            key = imageFile.uri,
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(imageInfo, resize, transformeds = null, extras = null)
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // char1 error
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        memoryCache.put(
+            key = requestContext.memoryCacheKey.replace("?_", "__"),
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(imageInfo, resize, transformeds = null, extras = null)
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // char2 error
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        memoryCache.put(
+            key = requestContext.memoryCacheKey.replace("?_", "??"),
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(imageInfo, resize, transformeds = null, extras = null)
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // thumbnail size within the error range
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        val thumbnailImage2 = createBitmap(thumbnailSize.width + 2, thumbnailSize.height).asImage()
+        memoryCache.put(
+            key = requestContext.memoryCacheKey,
+            value = ImageCacheValue(
+                image = thumbnailImage2,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = null,
+                    extras = null
                 )
-            }
-            requests1.plus(requests2).forEach { request ->
-                assertNull(
-                    inexactlyStateImage.getImage(sketch, request, null),
-                    request.toRequestContext(sketch).memoryCacheKey,
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = thumbnailImage2,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // thumbnail size without the error range
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        val thumbnailImage3 = createBitmap(thumbnailSize.width + 3, thumbnailSize.height).asImage()
+        memoryCache.put(
+            key = requestContext.memoryCacheKey,
+            value = ImageCacheValue(
+                image = thumbnailImage3,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = null,
+                    extras = null
                 )
-                assertNull(
-                    inexactlyStateImage1.getImage(sketch, request, null),
-                    request.toRequestContext(sketch).memoryCacheKey,
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // transformeds
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        memoryCache.put(
+            key = requestContext.memoryCacheKey,
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = listOf(createInSampledTransformed(4)),
+                    extras = null
                 )
-                assertNull(
-                    inexactlyStateImage2.getImage(sketch, request, null),
-                    request.toRequestContext(sketch).memoryCacheKey,
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = thumbnailImage,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+
+        // transformeds error
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        memoryCache.put(
+            key = requestContext.memoryCacheKey,
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = listOf(
+                        createInSampledTransformed(4),
+                        createResizeTransformed(resize)
+                    ),
+                    extras = null
                 )
-            }
+            )
+        )
+        assertEquals(expected = 1, actual = memoryCache.keys().size)
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
 
-            val testRequests1: suspend (Int) -> Unit = { loadIndex ->
-                memoryCache.clear()
-                assertEquals(0, memoryCache.keys().size)
-                sketch.enqueue(requests1[loadIndex]).job.await()
-                assertEquals(1, memoryCache.keys().size)
-                requests1.forEachIndexed { index, request ->
-                    if (index == loadIndex) {
-                        assertNotNull(
-                            memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    } else {
-                        assertNull(
-                            memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    }
-                }
-                requests2.forEach { request ->
-                    assertNull(
-                        memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                        request.toRequestContext(sketch).memoryCacheKey,
-                    )
-                }
-                requests1.forEach { request ->
-                    if (loadIndex == 0) {
-                        assertNotNull(
-                            inexactlyStateImage.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNotNull(
-                            inexactlyStateImage1.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNull(
-                            inexactlyStateImage2.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    } else {
-                        assertNull(
-                            inexactlyStateImage.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNull(
-                            inexactlyStateImage1.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNull(
-                            inexactlyStateImage2.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    }
-                }
-                requests2.forEach { request ->
-                    assertNull(inexactlyStateImage.getImage(sketch, request, null))
-                    if (loadIndex == 0) {
-                        assertNotNull(
-                            inexactlyStateImage1.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    } else {
-                        assertNull(
-                            inexactlyStateImage1.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    }
-                    assertNull(
-                        inexactlyStateImage2.getImage(sketch, request, null),
-                        request.toRequestContext(sketch).memoryCacheKey,
-                    )
-                }
-            }
-            testRequests1(0)
-            testRequests1(1)
-            testRequests1(2)
+        // maxMismatchCount
 
-            val testRequests2: suspend (Int) -> Unit = { loadIndex ->
-                memoryCache.clear()
-                assertEquals(0, memoryCache.keys().size)
-                sketch.enqueue(requests2[loadIndex]).job.await()
-                assertEquals(1, memoryCache.keys().size)
-                requests1.forEach { request ->
-                    assertNull(
-                        memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                        request.toRequestContext(sketch).memoryCacheKey,
-                    )
-                }
-                requests2.forEachIndexed { index, request ->
-                    if (index == loadIndex) {
-                        assertNotNull(
-                            memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    } else {
-                        assertNull(
-                            memoryCache[request.toRequestContext(sketch).memoryCacheKey],
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    }
-                }
-                requests1.forEach { request ->
-                    assertNull(inexactlyStateImage.getImage(sketch, request, null))
-                    assertNull(
-                        inexactlyStateImage1.getImage(sketch, request, null),
-                        request.toRequestContext(sketch).memoryCacheKey,
-                    )
-                    if (loadIndex == 0) {
-                        assertNotNull(
-                            inexactlyStateImage2.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    } else {
-                        assertNull(
-                            inexactlyStateImage2.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    }
-                }
-                requests2.forEach { request ->
-                    if (loadIndex == 0) {
-                        assertNotNull(
-                            inexactlyStateImage.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNull(
-                            inexactlyStateImage1.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNotNull(
-                            inexactlyStateImage2.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    } else {
-                        assertNull(
-                            inexactlyStateImage.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNull(
-                            inexactlyStateImage1.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                        assertNull(
-                            inexactlyStateImage2.getImage(sketch, request, null),
-                            request.toRequestContext(sketch).memoryCacheKey,
-                        )
-                    }
-                }
-            }
-            testRequests2(0)
-            testRequests2(1)
-            testRequests2(2)
-
-            memoryCache.clear()
-            assertEquals(0, memoryCache.keys().size)
-            sketch.enqueue(requests1[0]).job.await()
-            sketch.enqueue(requests2[0]).job.await()
-            assertEquals(2, memoryCache.keys().size)
-            requests1.forEachIndexed { index, request ->
-                if (index == 0) {
-                    assertNotNull(memoryCache[request.toRequestContext(sketch).memoryCacheKey])
-                } else {
-                    assertNull(memoryCache[request.toRequestContext(sketch).memoryCacheKey])
-                }
-            }
-            requests2.forEachIndexed { index, request ->
-                if (index == 0) {
-                    assertNotNull(memoryCache[request.toRequestContext(sketch).memoryCacheKey])
-                } else {
-                    assertNull(memoryCache[request.toRequestContext(sketch).memoryCacheKey])
-                }
-            }
-            requests1.forEach { request ->
-                assertNotNull(inexactlyStateImage.getImage(sketch, request, null))
-                assertNotNull(inexactlyStateImage1.getImage(sketch, request, null))
-                assertNotNull(inexactlyStateImage2.getImage(sketch, request, null))
-            }
-            requests2.forEach { request ->
-                assertNotNull(inexactlyStateImage.getImage(sketch, request, null))
-                assertNotNull(inexactlyStateImage1.getImage(sketch, request, null))
-                assertNotNull(inexactlyStateImage2.getImage(sketch, request, null))
-            }
-
-            memoryCache.clear()
-            assertEquals(0, memoryCache.keys().size)
-            sketch.enqueue(requests1[1]).job.await()
-            sketch.enqueue(requests1[2]).job.await()
-            sketch.enqueue(requests1[2].newRequest {
-                transformations(listOf(RoundedCornersTransformation()))
-            }).job.await()
-            sketch.enqueue(requests1[2].newRequest {
-                transformations(listOf(RotateTransformation(90)))
-            }).job.await()
-            assertNull(inexactlyStateImage.getImage(sketch, requests1[0], null))
-        }
+        memoryCache.clear()
+        assertEquals(0, memoryCache.keys().size)
+        memoryCache.put(
+            key = requestContext.memoryCacheKey + "&_test=1",
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = listOf(
+                        createInSampledTransformed(4),
+                        createResizeTransformed(resize)
+                    ),
+                    extras = null
+                )
+            )
+        )
+        memoryCache.put(
+            key = requestContext.memoryCacheKey + "&_test=2",
+            value = ImageCacheValue(
+                image = thumbnailImage3,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = null,
+                    extras = null
+                )
+            )
+        )
+        memoryCache.put(
+            key = requestContext.memoryCacheKey + "&_test=3",
+            value = ImageCacheValue(
+                image = thumbnailImage,
+                extras = newCacheValueExtras(
+                    imageInfo = imageInfo,
+                    resize = resize,
+                    transformeds = null,
+                    extras = null
+                )
+            )
+        )
+        assertEquals(3, memoryCache.keys().size)
+        assertEquals(
+            expected = thumbnailImage,
+            actual = ThumbnailMemoryCacheStateImage().getImage(sketch, request, null)
+        )
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage(maxMismatchCount = 0).getImage(
+                sketch,
+                request,
+                null
+            )
+        )
+        assertEquals(
+            expected = null,
+            actual = ThumbnailMemoryCacheStateImage(maxMismatchCount = 1).getImage(
+                sketch,
+                request,
+                null
+            )
+        )
+        assertEquals(
+            expected = thumbnailImage,
+            actual = ThumbnailMemoryCacheStateImage(maxMismatchCount = 2).getImage(
+                sketch,
+                request,
+                null
+            )
+        )
     }
 
     @Test
@@ -357,18 +374,24 @@ class ThumbnailMemoryCacheStateImageTest {
         val element4 =
             ThumbnailMemoryCacheStateImage(null, FakeStateImage(FakeImage(SketchSize(400, 400))))
         val element5 = ThumbnailMemoryCacheStateImage("uri1", null)
+        val element6 = ThumbnailMemoryCacheStateImage("uri1", null, 10)
 
         assertEquals(element1, element11)
         assertNotEquals(element1, element2)
         assertNotEquals(element1, element3)
         assertNotEquals(element1, element4)
         assertNotEquals(element1, element5)
+        assertNotEquals(element1, element6)
         assertNotEquals(element2, element3)
         assertNotEquals(element2, element4)
         assertNotEquals(element2, element5)
+        assertNotEquals(element2, element6)
         assertNotEquals(element3, element4)
         assertNotEquals(element3, element5)
+        assertNotEquals(element3, element6)
         assertNotEquals(element4, element5)
+        assertNotEquals(element4, element6)
+        assertNotEquals(element5, element6)
         assertNotEquals(element1, null as Any?)
         assertNotEquals(element1, Any())
 
@@ -377,12 +400,17 @@ class ThumbnailMemoryCacheStateImageTest {
         assertNotEquals(element1.hashCode(), element3.hashCode())
         assertNotEquals(element1.hashCode(), element4.hashCode())
         assertNotEquals(element1.hashCode(), element5.hashCode())
+        assertNotEquals(element1.hashCode(), element6.hashCode())
         assertNotEquals(element2.hashCode(), element3.hashCode())
         assertNotEquals(element2.hashCode(), element4.hashCode())
         assertNotEquals(element2.hashCode(), element5.hashCode())
+        assertNotEquals(element2.hashCode(), element6.hashCode())
         assertNotEquals(element3.hashCode(), element4.hashCode())
         assertNotEquals(element3.hashCode(), element5.hashCode())
+        assertNotEquals(element3.hashCode(), element6.hashCode())
         assertNotEquals(element4.hashCode(), element5.hashCode())
+        assertNotEquals(element4.hashCode(), element6.hashCode())
+        assertNotEquals(element5.hashCode(), element6.hashCode())
     }
 
     @Test
@@ -394,20 +422,20 @@ class ThumbnailMemoryCacheStateImageTest {
         val defaultImage = FakeStateImage()
         ThumbnailMemoryCacheStateImage(uri, defaultImage).apply {
             assertEquals(
-                "ThumbnailMemoryCacheStateImage(uri='$uri', defaultImage=$defaultImage)",
+                "ThumbnailMemoryCacheStateImage(uri='$uri', defaultImage=$defaultImage, maxMismatchCount=-1)",
                 toString()
             )
         }
         val defaultImage1 = FakeStateImage(FakeImage(SketchSize(200, 200)))
-        ThumbnailMemoryCacheStateImage(uri, defaultImage1).apply {
+        ThumbnailMemoryCacheStateImage(uri, defaultImage1, 10).apply {
             assertEquals(
-                "ThumbnailMemoryCacheStateImage(uri='$uri', defaultImage=$defaultImage1)",
+                "ThumbnailMemoryCacheStateImage(uri='$uri', defaultImage=$defaultImage1, maxMismatchCount=10)",
                 toString()
             )
         }
         ThumbnailMemoryCacheStateImage(null, null).apply {
             assertEquals(
-                "ThumbnailMemoryCacheStateImage(uri=null, defaultImage=null)",
+                "ThumbnailMemoryCacheStateImage(uri=null, defaultImage=null, maxMismatchCount=-1)",
                 toString()
             )
         }